+++ src/main/java/org/jabref/gui/preferences/file/FileTab.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.control.ToggleGroup?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-<?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
+
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.file.FileTab">
@@ -36,9 +36,6 @@
         <ComboBox fx:id="newLineSeparator" prefWidth="120.0"/>
     </HBox>
     <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
-
-    <Label styleClass="sectionHeader" text="%Export sort order"/>
-    <SaveOrderConfigPanel fx:id="exportOrderPanel"/>
 
     <Label styleClass="sectionHeader" text="%Autosave"/>
     <HBox alignment="CENTER_LEFT" spacing="10.0">
