+++ src/main/java/org/jabref/gui/preferences/importexport/ImportExportTab.fxml
@@ -5,6 +5,8 @@
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
+<?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
+
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.importexport.ImportExportTab">
@@ -18,4 +20,7 @@
         <CheckBox fx:id="useCustomDOI" text="%Use custom DOI base URI for article access"/>
         <TextField fx:id="useCustomDOIName" HBox.hgrow="ALWAYS"/>
     </HBox>
+
+    <Label styleClass="sectionHeader" text="%Export sort order"/>
+    <SaveOrderConfigPanel fx:id="exportOrderPanel"/>
 </fx:root>
