+++ src/main/java/org/jabref/gui/preferences/importexport/ImportExportTabViewModel.java
@@ -1,13 +1,26 @@
 package org.jabref.gui.preferences.importexport;
 
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.ListProperty;
 import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleListProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
+import javafx.collections.FXCollections;
 
+import org.jabref.gui.commonfxcontrols.SortCriterionViewModel;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.logic.importer.importsettings.ImportSettingsPreferences;
 import org.jabref.logic.preferences.DOIPreferences;
+import org.jabref.model.entry.field.Field;
+import org.jabref.model.entry.field.FieldFactory;
+import org.jabref.model.metadata.SaveOrderConfig;
 import org.jabref.preferences.PreferencesService;
 
 public class ImportExportTabViewModel implements PreferenceTabViewModel {
@@ -17,15 +30,23 @@ public class ImportExportTabViewModel implements PreferenceTabViewModel {
     private final BooleanProperty useCustomDOIProperty = new SimpleBooleanProperty();
     private final StringProperty useCustomDOINameProperty = new SimpleStringProperty("");
 
-    private final PreferencesService preferencesService;
+    // SaveOrderConfigPanel
+    private final BooleanProperty exportInOriginalProperty = new SimpleBooleanProperty();
+    private final BooleanProperty exportInTableOrderProperty = new SimpleBooleanProperty();
+    private final BooleanProperty exportInSpecifiedOrderProperty = new SimpleBooleanProperty();
+    private final ListProperty<Field> sortableFieldsProperty = new SimpleListProperty<>(FXCollections.observableArrayList());
+    private final ListProperty<SortCriterionViewModel> sortCriteriaProperty = new SimpleListProperty<>(FXCollections.observableArrayList(new LinkedList<>()));
 
+    private final PreferencesService preferencesService;
     private final DOIPreferences initialDOIPreferences;
     private final ImportSettingsPreferences initialImportSettingsPreferences;
+    private final SaveOrderConfig initialExportOrder;
 
     public ImportExportTabViewModel(PreferencesService preferencesService) {
         this.preferencesService = preferencesService;
         this.initialImportSettingsPreferences = preferencesService.getImportSettingsPreferences();
         this.initialDOIPreferences = preferencesService.getDOIPreferences();
+        this.initialExportOrder = preferencesService.getExportSaveOrder();
     }
 
     @Override
@@ -33,6 +54,20 @@ public class ImportExportTabViewModel implements PreferenceTabViewModel {
         generateKeyOnImportProperty.setValue(initialImportSettingsPreferences.generateNewKeyOnImport());
         useCustomDOIProperty.setValue(initialDOIPreferences.isUseCustom());
         useCustomDOINameProperty.setValue(initialDOIPreferences.getDefaultBaseURI());
+
+        switch (initialExportOrder.getOrderType()) {
+            case SPECIFIED -> exportInSpecifiedOrderProperty.setValue(true);
+            case ORIGINAL -> exportInOriginalProperty.setValue(true);
+            case TABLE -> exportInTableOrderProperty.setValue(true);
+        }
+
+        List<Field> fieldNames = new ArrayList<>(FieldFactory.getCommonFields());
+        fieldNames.sort(Comparator.comparing(Field::getDisplayName));
+
+        sortableFieldsProperty.addAll(fieldNames);
+        sortCriteriaProperty.addAll(initialExportOrder.getSortCriteria().stream()
+                                                      .map(SortCriterionViewModel::new)
+                                                      .collect(Collectors.toList()));
     }
 
     @Override
@@ -43,6 +78,12 @@ public class ImportExportTabViewModel implements PreferenceTabViewModel {
         preferencesService.storeDOIPreferences(new DOIPreferences(
                 useCustomDOIProperty.getValue(),
                 useCustomDOINameProperty.getValue().trim()));
+
+        SaveOrderConfig newSaveOrderConfig = new SaveOrderConfig(
+                SaveOrderConfig.OrderType.fromBooleans(exportInSpecifiedOrderProperty.getValue(), exportInTableOrderProperty.getValue()),
+                new LinkedList<>(sortCriteriaProperty.stream().map(SortCriterionViewModel::getCriterion).toList()));
+        preferencesService.storeExportSaveOrder(newSaveOrderConfig);
+
     }
 
     public BooleanProperty generateKeyOnImportProperty() {
@@ -55,5 +96,27 @@ public class ImportExportTabViewModel implements PreferenceTabViewModel {
 
     public StringProperty useCustomDOINameProperty() {
         return this.useCustomDOINameProperty;
+    }
+
+    // SaveOrderConfigPanel
+
+    public BooleanProperty saveInOriginalProperty() {
+        return exportInOriginalProperty;
+    }
+
+    public BooleanProperty saveInTableOrderProperty() {
+        return exportInTableOrderProperty;
+    }
+
+    public BooleanProperty saveInSpecifiedOrderProperty() {
+        return exportInSpecifiedOrderProperty;
+    }
+
+    public ListProperty<Field> sortableFieldsProperty() {
+        return sortableFieldsProperty;
+    }
+
+    public ListProperty<SortCriterionViewModel> sortCriteriaProperty() {
+        return sortCriteriaProperty;
     }
 }
